Move catalog tests into folder.
