+++ spec/src/modules/catalog/catalog-groups.js
@@ -7,8 +7,8 @@ const sinonChai = require('sinon-chai');
 const nodeFetch = require('node-fetch').default;
 const cloneDeep = require('lodash.clonedeep');
 const { v4: uuidv4 } = require('uuid');
-const ConstructorIO = require('../../../test/constructorio'); // eslint-disable-line import/extensions
-const helpers = require('../../mocha.helpers');
+const ConstructorIO = require('../../../../test/constructorio'); // eslint-disable-line import/extensions
+const helpers = require('../../../mocha.helpers');
 
 chai.use(chaiAsPromised);
 chai.use(sinonChai);
