+++ spec/src/modules/catalog/catalog-items.js
@@ -7,15 +7,14 @@ const sinonChai = require('sinon-chai');
 const nodeFetch = require('node-fetch').default;
 const cloneDeep = require('lodash.clonedeep');
 const { v4: uuidv4 } = require('uuid');
-const ConstructorIO = require('../../../test/constructorio'); // eslint-disable-line import/extensions
-const helpers = require('../../mocha.helpers');
+const ConstructorIO = require('../../../../test/constructorio'); // eslint-disable-line import/extensions
+const helpers = require('../../../mocha.helpers');
 
 chai.use(chaiAsPromised);
 chai.use(sinonChai);
 dotenv.config();
 
 const itemsToCleanup = [];
-const variationsToCleanup = [];
 const sendTimeout = 300;
 const testApiKey = process.env.TEST_API_KEY;
 const testApiToken = process.env.TEST_API_TOKEN;
@@ -32,26 +31,6 @@ function createMockItem() {
     id: uuid,
     data: {
       facets: { color: ['blue', 'red'] },
-      brand: 'abc',
-      url: 'https://constructor.io/products/',
-      image_url: 'https://constructor.io/products/',
-      complexMetadataField: {
-        key1: 'val1',
-        key2: 'val2',
-      },
-    },
-  };
-}
-
-function createMockVariation(itemId) {
-  const uuid = uuidv4();
-
-  return {
-    id: uuid,
-    item_id: itemId,
-    name: `product-${uuid}`,
-    data: {
-      facets: { color: ['blue'] },
       brand: 'abc',
       url: 'https://constructor.io/products/',
       image_url: 'https://constructor.io/products/',
@@ -81,14 +60,12 @@ describe('ConstructorIO - Catalog', () => {
     setTimeout(done, sendTimeout);
   });
 
-  describe('Variations', () => {
-    const item = createMockItem();
-
-    describe('createOrReplaceVariations', () => {
-      const variations = [
-        createMockVariation(item.id),
-        createMockVariation(item.id),
-        createMockVariation(item.id),
+  describe('Items', () => {
+    describe('createOrReplaceItems', () => {
+      const items = [
+        createMockItem(),
+        createMockItem(),
+        createMockItem(),
       ];
       const optionalParameters = {
         section: 'Products',
@@ -96,13 +73,13 @@ describe('ConstructorIO - Catalog', () => {
         notificationEmail: 'test@constructor.io',
       };
 
-      it('Should resolve when adding multiple variations', (done) => {
+      it('Should resolve when adding multiple items', (done) => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
 
-        catalog.createOrReplaceVariations({ variations }).then(() => {
+        catalog.createOrReplaceItems({ items }).then(() => {
           const requestedUrlParams = helpers.extractUrlParamsFromFetch(fetchSpy);
 
           expect(fetchSpy).to.have.been.called;
@@ -110,17 +87,16 @@ describe('ConstructorIO - Catalog', () => {
           done();
         });
 
-        itemsToCleanup.push(item);
-        variationsToCleanup.push(...variations);
+        itemsToCleanup.push(...items);
       });
 
-      it('Should resolve when adding multiple variations and supplying optional parameters', (done) => {
+      it('Should resolve when adding multiple items and supplying optional parameters', (done) => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
 
-        catalog.createOrReplaceVariations({ variations, ...optionalParameters }).then(() => {
+        catalog.createOrReplaceItems({ items, ...optionalParameters }).then(() => {
           const requestedUrlParams = helpers.extractUrlParamsFromFetch(fetchSpy);
 
           expect(requestedUrlParams).to.have.property('section').to.equal(optionalParameters.section);
@@ -129,11 +105,10 @@ describe('ConstructorIO - Catalog', () => {
           done();
         });
 
-        itemsToCleanup.push(item);
-        variationsToCleanup.push(...variations);
+        itemsToCleanup.push(...items);
       });
 
-      it('Should return error when no variations are provided', () => {
+      it('Should return error when no items are provided', () => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
@@ -142,16 +117,16 @@ describe('ConstructorIO - Catalog', () => {
         return expect(catalog.createOrReplaceItems()).to.eventually.be.rejected;
       });
 
-      it('Should return error when invalid variations are provided', () => {
+      it('Should return error when invalid items are provided', () => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
 
-        return expect(catalog.createOrReplaceItems({ variations: 'foo' })).to.eventually.be.rejected;
+        return expect(catalog.createOrReplaceItems({ items: 'foo' })).to.eventually.be.rejected;
       });
 
-      it('Should return error when adding multiple variations with an invalid API key', () => {
+      it('Should return error when adding multiple items with an invalid API key', () => {
         const invalidOptions = cloneDeep(validOptions);
 
         invalidOptions.apiKey = 'abc123';
@@ -161,10 +136,10 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        return expect(catalog.createOrReplaceVariations({ variations })).to.eventually.be.rejected;
+        return expect(catalog.createOrReplaceItems({ items })).to.eventually.be.rejected;
       });
 
-      it('Should return error when adding multiple variations with an invalid API token', () => {
+      it('Should return error when adding multiple items with an invalid API token', () => {
         const invalidOptions = cloneDeep(validOptions);
 
         invalidOptions.apiToken = 'foo987';
@@ -174,13 +149,13 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        return expect(catalog.createOrReplaceVariations({ variations })).to.eventually.be.rejected;
+        return expect(catalog.createOrReplaceItems({ items })).to.eventually.be.rejected;
       });
 
       it('Should be rejected when network request timeout is provided and reached', () => {
         const { catalog } = new ConstructorIO(validOptions);
 
-        return expect(catalog.createOrReplaceVariations({ variations: [createMockVariation(item.id)] }, { timeout: 10 })).to.eventually.be.rejectedWith('The user aborted a request.');
+        return expect(catalog.createOrReplaceItems({ items: [createMockItem()] }, { timeout: 10 })).to.eventually.be.rejectedWith('The user aborted a request.');
       });
 
       it('Should be rejected when global network request timeout is provided and reached', () => {
@@ -189,13 +164,13 @@ describe('ConstructorIO - Catalog', () => {
           networkParameters: { timeout: 20 },
         });
 
-        return expect(catalog.createOrReplaceVariations({ variations: [createMockVariation(item.id)] })).to.eventually.be.rejectedWith('The user aborted a request.');
+        return expect(catalog.createOrReplaceItems({ items: [createMockItem()] })).to.eventually.be.rejectedWith('The user aborted a request.');
       });
     });
 
-    describe('updateVariations', () => {
-      const variations = [createMockVariation(item.id)];
-      const updatedVariations = [{ ...variations[0], name: 'Updated Variation Name' }];
+    describe('updateItems', () => {
+      const items = [createMockItem()];
+      const updatedItems = [{ ...items[0], name: 'Updated Item Name' }];
       const optionalParameters = {
         section: 'Products',
         force: true,
@@ -208,17 +183,17 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        catalog.createOrReplaceVariations({ variations }).then(done);
-        variationsToCleanup.push(...updatedVariations);
+        catalog.createOrReplaceItems({ items }).then(done);
+        itemsToCleanup.push(...items);
       });
 
-      it('Should resolve when updating multiple variations', (done) => {
+      it('Should resolve when updating multiple items', (done) => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
 
-        catalog.updateVariations({ variations: updatedVariations }).then(() => {
+        catalog.updateItems({ items: updatedItems }).then(() => {
           const requestedUrlParams = helpers.extractUrlParamsFromFetch(fetchSpy);
 
           expect(fetchSpy).to.have.been.called;
@@ -227,13 +202,13 @@ describe('ConstructorIO - Catalog', () => {
         });
       });
 
-      it('Should resolve when updating multiple variations and supplying optional parameters', (done) => {
+      it('Should resolve when updating multiple items and supplying optional parameters', (done) => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
 
-        catalog.updateVariations({ variations: updatedVariations, ...optionalParameters }).then(() => {
+        catalog.updateItems({ items: updatedItems, ...optionalParameters }).then(() => {
           const requestedUrlParams = helpers.extractUrlParamsFromFetch(fetchSpy);
 
           expect(requestedUrlParams).to.have.property('section').to.equal(optionalParameters.section);
@@ -243,25 +218,7 @@ describe('ConstructorIO - Catalog', () => {
         });
       });
 
-      it('Should return error when no variations are provided', () => {
-        const { catalog } = new ConstructorIO({
-          ...validOptions,
-          fetch: fetchSpy,
-        });
-
-        return expect(catalog.updateVariations()).to.eventually.be.rejected;
-      });
-
-      it('Should return error when invalid variations are provided', () => {
-        const { catalog } = new ConstructorIO({
-          ...validOptions,
-          fetch: fetchSpy,
-        });
-
-        return expect(catalog.updateVariations({ variation: 'foo' })).to.eventually.be.rejected;
-      });
-
-      it('Should return error when updating variations with an invalid API key', () => {
+      it('Should return error when no items are provided', () => {
         const invalidOptions = cloneDeep(validOptions);
 
         invalidOptions.apiKey = 'abc123';
@@ -271,10 +228,36 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        return expect(catalog.updateVariations({ variations: updatedVariations })).to.eventually.be.rejected;
+        return expect(catalog.updateItems()).to.eventually.be.rejected;
       });
 
-      it('Should return error when updating variations with an invalid API token', () => {
+      it('Should return error when invalid items are provided', () => {
+        const invalidOptions = cloneDeep(validOptions);
+
+        invalidOptions.apiKey = 'abc123';
+
+        const { catalog } = new ConstructorIO({
+          ...invalidOptions,
+          fetch: fetchSpy,
+        });
+
+        return expect(catalog.updateItems({ items: 'foo' })).to.eventually.be.rejected;
+      });
+
+      it('Should return error when updating items with an invalid API key', () => {
+        const invalidOptions = cloneDeep(validOptions);
+
+        invalidOptions.apiKey = 'abc123';
+
+        const { catalog } = new ConstructorIO({
+          ...invalidOptions,
+          fetch: fetchSpy,
+        });
+
+        return expect(catalog.updateItems({ items: updatedItems })).to.eventually.be.rejected;
+      });
+
+      it('Should return error when updating items with an invalid API token', () => {
         const invalidOptions = cloneDeep(validOptions);
 
         invalidOptions.apiToken = 'foo987';
@@ -284,13 +267,13 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        return expect(catalog.updateVariations({ variations: updatedVariations })).to.eventually.be.rejected;
+        return expect(catalog.updateItems({ items: updatedItems })).to.eventually.be.rejected;
       });
 
       it('Should be rejected when network request timeout is provided and reached', () => {
         const { catalog } = new ConstructorIO(validOptions);
 
-        return expect(catalog.updateVariations({ variations: updatedVariations }, { timeout: 10 })).to.eventually.be.rejectedWith('The user aborted a request.');
+        return expect(catalog.updateItems({ items: updatedItems }, { timeout: 10 })).to.eventually.be.rejectedWith('The user aborted a request.');
       });
 
       it('Should be rejected when network global request timeout is provided and reached', () => {
@@ -299,20 +282,20 @@ describe('ConstructorIO - Catalog', () => {
           networkParameters: { timeout: 20 },
         });
 
-        return expect(catalog.updateVariations({ variations: updatedVariations })).to.eventually.be.rejectedWith('The user aborted a request.');
+        return expect(catalog.updateItems({ items: updatedItems })).to.eventually.be.rejectedWith('The user aborted a request.');
       });
     });
 
-    describe('deleteVariations', () => {
-      const variations = [
-        createMockVariation(item.id),
-        createMockVariation(item.id),
-        createMockVariation(item.id),
+    describe('deleteItems', () => {
+      const items = [
+        createMockItem(),
+        createMockItem(),
+        createMockItem(),
       ];
-      const variationsDoNotExist = [
-        createMockVariation(item.id),
-        createMockVariation(item.id),
-        createMockVariation(item.id),
+      const itemsDoNotExist = [
+        createMockItem(),
+        createMockItem(),
+        createMockItem(),
       ];
 
       before((done) => {
@@ -321,75 +304,47 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        catalog.createOrReplaceVariations({ variations, section: 'Products' }).then(done);
+        catalog.createOrReplaceItems({ items }).then(done);
+        itemsToCleanup.push(...items);
       });
 
-      it('Should resolve when removing multiple variations', (done) => {
+      it('Should resolve when removing multiple items', (done) => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
 
-        variations.push(...variationsToCleanup);
-        catalog.deleteVariations({ variations: variations.map((variation) => ({ id: variation.id })), section: 'Products' }).then(() => {
+        catalog.deleteItems({ items: items.map((item) => ({ id: item.id })) }).then(() => {
           const requestedUrlParams = helpers.extractUrlParamsFromFetch(fetchSpy);
 
           expect(fetchSpy).to.have.been.called;
           expect(requestedUrlParams).to.have.property('key');
-
-          catalog.deleteItems({ items: itemsToCleanup.map((variation) => ({ id: variation.id })), section: 'Products' }).then(done);
+          done();
         });
       });
 
-      it('Should resolve when removing multiple variations and supplying optional parameters', (done) => {
+      it('Should resolve when removing multiple items and supplying optional parameters', (done) => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
         const optionalParameters = {
+          section: 'Products',
           force: true,
           notificationEmail: 'test@constructor.io',
         };
 
-        variations.push(...variationsToCleanup);
-        catalog.deleteVariations({ variations: variations.map((variation) => ({ id: variation.id })), section: 'Products', ...optionalParameters }).then(() => {
+        catalog.deleteItems({ items: items.map((item) => ({ id: item.id })), ...optionalParameters }).then(() => {
           const requestedUrlParams = helpers.extractUrlParamsFromFetch(fetchSpy);
 
+          expect(requestedUrlParams).to.have.property('section').to.equal(optionalParameters.section);
           expect(requestedUrlParams).to.have.property('force').to.equal(optionalParameters.force.toString());
           expect(requestedUrlParams).to.have.property('notification_email').to.equal(optionalParameters.notificationEmail);
-
-          catalog.deleteItems({ items: itemsToCleanup.map((variation) => ({ id: variation.id })), section: 'Products' }).then(done);
+          done();
         });
       });
 
-      it('Should return error when removing variations that do not exist', () => {
-        const { catalog } = new ConstructorIO({
-          ...validOptions,
-          fetch: fetchSpy,
-        });
-
-        return expect(catalog.deleteVariations({ variations: variationsDoNotExist, section: 'Products' })).to.eventually.be.rejected;
-      });
-
-      it('Should return error when no variations are provided', () => {
-        const { catalog } = new ConstructorIO({
-          ...validOptions,
-          fetch: fetchSpy,
-        });
-
-        return expect(catalog.deleteVariations()).to.eventually.be.rejected;
-      });
-
-      it('Should return error when invalid variations are provided', () => {
-        const { catalog } = new ConstructorIO({
-          ...validOptions,
-          fetch: fetchSpy,
-        });
-
-        return expect(catalog.deleteVariations({ variations: 'foo' })).to.eventually.be.rejected;
-      });
-
-      it('Should return error when removing variations with an invalid API key', () => {
+      it('Should return error when no items are provided', () => {
         const invalidOptions = cloneDeep(validOptions);
 
         invalidOptions.apiKey = 'abc123';
@@ -399,10 +354,47 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        return expect(catalog.deleteVariations({ variations, section: 'Products' })).to.eventually.be.rejected;
+        return expect(catalog.deleteItems()).to.eventually.be.rejected;
       });
 
-      it('Should return error when removing variations with an invalid API token', () => {
+      it('Should return error when invalid items are provided', () => {
+        const invalidOptions = cloneDeep(validOptions);
+
+        invalidOptions.apiKey = 'abc123';
+
+        const { catalog } = new ConstructorIO({
+          ...invalidOptions,
+          fetch: fetchSpy,
+        });
+
+        return expect(catalog.deleteItems({ items: 'foo' })).to.eventually.be.rejected;
+      });
+
+      it('Should return error when removing items that do not exist', () => {
+        const invalidOptions = cloneDeep(validOptions);
+
+        const { catalog } = new ConstructorIO({
+          ...invalidOptions,
+          fetch: fetchSpy,
+        });
+
+        return expect(catalog.deleteItems({ items: itemsDoNotExist })).to.eventually.be.rejected;
+      });
+
+      it('Should return error when removing items with an invalid API key', () => {
+        const invalidOptions = cloneDeep(validOptions);
+
+        invalidOptions.apiKey = 'abc123';
+
+        const { catalog } = new ConstructorIO({
+          ...invalidOptions,
+          fetch: fetchSpy,
+        });
+
+        return expect(catalog.deleteItems({ items })).to.eventually.be.rejected;
+      });
+
+      it('Should return error when removing items with an invalid API token', () => {
         const invalidOptions = cloneDeep(validOptions);
 
         invalidOptions.apiToken = 'foo987';
@@ -412,13 +404,13 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        return expect(catalog.deleteVariations({ variations, section: 'Products' })).to.eventually.be.rejected;
+        return expect(catalog.deleteItems({ items })).to.eventually.be.rejected;
       });
 
       it('Should be rejected when network request timeout is provided and reached', () => {
         const { catalog } = new ConstructorIO(validOptions);
 
-        return expect(catalog.deleteVariations({ variations, section: 'Products' }, { timeout: 10 })).to.eventually.be.rejectedWith('The user aborted a request.');
+        return expect(catalog.deleteItems({ items }, { timeout: 10 })).to.eventually.be.rejectedWith('The user aborted a request.');
       });
 
       it('Should be rejected when global network request timeout is provided and reached', () => {
@@ -427,12 +419,12 @@ describe('ConstructorIO - Catalog', () => {
           networkParameters: { timeout: 20 },
         });
 
-        return expect(catalog.deleteVariations({ variations, section: 'Products' })).to.eventually.be.rejectedWith('The user aborted a request.');
+        return expect(catalog.deleteItems({ items })).to.eventually.be.rejectedWith('The user aborted a request.');
       });
     });
 
-    describe('retrieveVariations', () => {
-      const variations = [createMockVariation(item.id), createMockVariation(item.id), createMockVariation(item.id)];
+    describe('retrieveItems', () => {
+      const mockItems = [createMockItem(), createMockItem(), createMockItem()];
 
       before((done) => {
         const { catalog } = new ConstructorIO({
@@ -440,68 +432,43 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        catalog.createOrReplaceVariations({ variations, section: 'Products' }).then((done));
-        variationsToCleanup.push(...variations);
+        catalog.createOrReplaceItems({ items: mockItems }).then((done));
+        itemsToCleanup.push(...mockItems);
       });
 
-      it('Should return a response when getting variations', (done) => {
+      it('Should return a response when getting items', (done) => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
 
-        catalog.retrieveVariations().then((res) => {
+        catalog.retrieveItems().then((res) => {
           const requestedUrlParams = helpers.extractUrlParamsFromFetch(fetchSpy);
 
-          expect(res).to.have.property('variations').to.be.an('array');
+          expect(res).to.have.property('items').to.be.an('array');
           expect(fetchSpy).to.have.been.called;
           expect(requestedUrlParams).to.have.property('key');
           done();
         });
       });
 
-      it('Should return a response when getting variations by id', (done) => {
-        const { catalog } = new ConstructorIO({
-          ...validOptions,
-          fetch: fetchSpy,
-        });
-
-        catalog.retrieveVariations({ ids: variations.map(((variation) => variation.id)) }).then((res) => {
-          expect(res).to.have.property('variations').to.be.an('array');
-          done();
-        });
-      });
-
-      it('Should return a response when getting variations by item id', (done) => {
-        const { catalog } = new ConstructorIO({
-          ...validOptions,
-          fetch: fetchSpy,
-        });
-
-        catalog.retrieveVariations({ itemId: item.id }).then((res) => {
-          const requestedUrlParams = helpers.extractUrlParamsFromFetch(fetchSpy);
-
-          expect(res).to.have.property('variations').to.be.an('array');
-          expect(requestedUrlParams).to.have.property('item_id').to.be.an('string').to.equal(item.id);
-          done();
-        });
-      });
-
-      it('Should return a response when getting variations by section with optional parameters', (done) => {
+      it('Should return a response when getting items with optional parameters', (done) => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
         const optionalParameters = {
           section: 'Products',
+          ids: mockItems.map(((item) => item.id)),
           numResultsPerPage: 10,
           page: 1,
         };
 
-        catalog.retrieveVariations(optionalParameters).then((res) => {
+        catalog.retrieveItems(optionalParameters).then((res) => {
           const requestedUrlParams = helpers.extractUrlParamsFromFetch(fetchSpy);
 
-          expect(res).to.have.property('variations').to.be.an('array');
+          expect(res).to.have.property('items').to.be.an('array');
+          expect(requestedUrlParams).to.have.property('id').to.be.an('array').to.deep.equal(optionalParameters.ids);
           expect(requestedUrlParams).to.have.property('section').to.equal(optionalParameters.section);
           expect(requestedUrlParams).to.have.property('num_results_per_page').to.equal(optionalParameters.numResultsPerPage.toString());
           expect(requestedUrlParams).to.have.property('page').to.equal(optionalParameters.page.toString());
@@ -509,26 +476,14 @@ describe('ConstructorIO - Catalog', () => {
         });
       });
 
-      it('Should return variations of length 0 when getting variation by id that does not exist', (done) => {
+      it('Should return items of length 0 when getting item by id that does not exist', (done) => {
         const { catalog } = new ConstructorIO({
           ...validOptions,
           fetch: fetchSpy,
         });
 
-        catalog.retrieveVariations({ ids: [uuidv4()] }).then((res) => {
-          expect(res).to.have.property('variations').to.be.an('array').that.is.empty;
-          done();
-        });
-      });
-
-      it('Should return variations of length 0 when getting variations by item id that does not exist', (done) => {
-        const { catalog } = new ConstructorIO({
-          ...validOptions,
-          fetch: fetchSpy,
-        });
-
-        catalog.retrieveVariations({ itemId: 'abc' }).then((res) => {
-          expect(res).to.have.property('variations').to.be.an('array').that.is.empty;
+        catalog.retrieveItems({ ids: [uuidv4()], section: 'Products' }).then((res) => {
+          expect(res).to.have.property('items').to.be.an('array').that.is.empty;
           done();
         });
       });
@@ -543,7 +498,7 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        return expect(catalog.retrieveVariations({ section: 'Products' })).to.eventually.be.rejected;
+        return expect(catalog.retrieveItems()).to.eventually.be.rejected;
       });
 
       it('Should return error when retrieving an item with an invalid API token', () => {
@@ -556,13 +511,13 @@ describe('ConstructorIO - Catalog', () => {
           fetch: fetchSpy,
         });
 
-        return expect(catalog.retrieveVariations({ section: 'Products' })).to.eventually.be.rejected;
+        return expect(catalog.retrieveItems()).to.eventually.be.rejected;
       });
 
       it('Should be rejected when network request timeout is provided and reached', () => {
         const { catalog } = new ConstructorIO(validOptions);
 
-        return expect(catalog.retrieveVariations({ section: 'Products' }, { timeout: 10 })).to.eventually.be.rejectedWith('The user aborted a request.');
+        return expect(catalog.retrieveItems({}, { timeout: 10 })).to.eventually.be.rejectedWith('The user aborted a request.');
       });
 
       it('Should be rejected when global network request timeout is provided and reached', () => {
@@ -571,7 +526,7 @@ describe('ConstructorIO - Catalog', () => {
           networkParameters: { timeout: 20 },
         });
 
-        return expect(catalog.retrieveVariations({ section: 'Products' })).to.eventually.be.rejectedWith('The user aborted a request.');
+        return expect(catalog.retrieveItems()).to.eventually.be.rejectedWith('The user aborted a request.');
       });
     });
   });
